Profile Resolvers & Update Experience
Create resolvers in order to first load the profile user and after the
profile data to avoid loading data for the current user in other
profiles. Implement the Angular functionality for the user to be able to
add, update and delete his experience his profile. The user must be able
to edit only his profile. Implement a method myProfile() to check that
at the components that need it. Create the java implementation for the
update of experience and change the visible from enum to integer.