Profile Update functionality - Education & Modal style fix
Education creation, update and deletion for each user. Update the
education list using subject. Removed the @NotNull from the Education
object to allow empty values in endDate (for Present education). Change
education's dto field from id to educationId to be consistent with the
other profile attributes. Fix the modal in all the profile's modals by
styling them and make them cover the whole modal and not the half.