Profile Post - View and Delete
Create post object and basic functionality in Angular. Get all users
posts and diplay them at his profile in the activity section. As the
user can create a lot of post it is better the post section to be
scrollable. The user can delete a post just by clicking the bin right
next to the post. Make all the profile's attributes equal size.