Profile Post - Edit post And Show Likes
The user should be able to update one of his posts through his profile
and also to be able to see the number of likes he got. Rename the
RequestPostDto to PostRequestDto for consistency and remove the date and
userId fields as they are unnecessary. Adjust PostService to set the
Date only on create and not in update.