Profile Post - Like functionality
Change Post.csv likes to 1 in order to be valid.
Java: rename the package from Likes to likes, create unlike endpoint and
service. Delete the LikeRequestDto as it is not used. Change userDto
imgPath to imagePath to match with userSimpleDto.

Angular: create the like object and service. Change the imgPath to
imagePath. Fix constructor argument type in skill model. Remove
unnecessary html code in newsfeed component. In post-list add the like
and unlike functionality. Create a like modal that shows all the people
who liked the posts. Change navbar to use the user's image in top right.