Add new connections
Create "add connection" functionality when user visits a profile that he
is not friend with. Add a resolver at profile for the logged in logged
user to get his friends in order to check if he is friend with one of
them or not, and one resolver at the user's profile we are at the moment
in order to avoid the wrong load call from navigating through different
profiles.