Chating and Messages
Angular:
Create Messaging interface in the application. One component to show
user's chat overview and one child component to show the selected
conversation. Add one resolver in the child in order to load the
messages properly between the navigations. Add the redirect to the chat
from the navbar and from user's profile.

Java:
Change the allow-origin in CORS to * in order to accept https too.
Change MessageDto to the fields that we need (message, sendBy, date).
Change UserChatDto to return a list of messages and basic details for
the 2nd participant of the conversation. Remove unnecessary prints from
Authentication Facade. Adjust the new Message Dtos' changes in the
service and get the conversation's message by ascending date.